refactor(ai-context): graduate context substrate out of studio (Phase 0)
Move the agentic-metadata engine (context-dirty, context-refresh, metadata,
gen-lock, context-spend, source-resolver, tokens + substrate types) from
extensions/studio/server/ to lib/domain/ai-context/ — it is per-ContentNode
content-graph infrastructure now consumed beyond the studio extension.
Code moves, identifiers stay (table names, route ids, settings keys).

Also: replace retired claude-haiku-3-5 defaultSuggestion (retired
2026-02-19) with claude-haiku-4-5 across six feature routes.

Plan: docs/notes-feature/work-tracking/FOLDER-CONTEXT-CAPSULE-PLAN.md
Gates: typecheck clean; lint 0 errors (151 pre-existing warnings < 175).
