+++ lib/domain/ai-context/context-dirty.ts
