+++ lib/domain/ai-context/context-spend.ts
