+++ lib/domain/ai-context/gen-lock.ts
